course activity now displays Users within that course aswell

+++ part2/app/src/main/java/com/example/part2/StudentCourses.java
@@ -10,20 +10,20 @@ public class StudentCourses {
 
 
     @ColumnInfo(name = "student_id")
-    private long student_id;
+    private int student_id;
     @ColumnInfo(name = "course_id")
-    private long course_id;
+    private int course_id;
 
 
-    public long getCourse_id() {
+    public int getCourse_id() {
         return course_id;
     }
 
-    public long getStudent_id() {
+    public int getStudent_id() {
         return student_id;
     }
 
-    public StudentCourses(long student_id, long course_id) {
+    public StudentCourses(int student_id, int course_id) {
         this.student_id = student_id;
         this.course_id = course_id;
     }
